pyproject.toml
[tool.poetry]
name = "hatchet-sdk"
version = "0.34.0a5"
description = ""
authors = ["Alexander Belanger <alexander@hatchet.run>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.64.1"
python-dotenv = "^1.0.0"
protobuf = "^4.25.2"
pyyaml = "^6.0.1"
grpcio-tools = "^1.60.0"
loguru = "^0.7.2"
pydantic = "^2.6.3"
python-dateutil = "^2.9.0.post0"
urllib3 = "^2.2.1"
aiostream = "^0.5.2"
nest-asyncio = "^1.6.0"

rizaio = { version = "^0.1.0a8", allow-prereleases = true }
[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.8"
psutil = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true

[tool.isort]
profile = "black"
known_third_party = [
  "grpcio",
  "grpcio_tools",
  "loguru",
  "protobuf",
  "pydantic",
  "python_dotenv",
  "python_dateutil",
  "pyyaml",
  "urllib3",
]

[tool.poetry.scripts]
api = "examples.api.api:main"
async = "examples.async.worker:main"
fanout = "examples.fanout.worker:main"
riza = "examples.riza.worker:main"
cancellation = "examples.cancellation.worker:main"
concurrency_limit = "examples.concurrency_limit.worker:main"
concurrency_limit_rr = "examples.concurrency_limit_rr.worker:main"
dag = "examples.dag.worker:main"
delayed = "examples.delayed.worker:main"
events = "examples.events.worker:main"
logger = "examples.logger.worker:main"
manual_trigger = "examples.manual_trigger.worker:main"
on_failure = "examples.on_failure.worker:main"
programatic_replay = "examples.programatic_replay.worker:main"
rate_limit = "examples.rate_limit.worker:main"
simple = "examples.simple.worker:main"
timeout = "examples.timeout.worker:main"
blocked = "examples.blocked_async.worker:main"
